requirements.txt
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
plotly>=5.10.0
spectral>=0.22.4
tifffile>=2022.0.0
h5py>=3.0.0
PyGithub>=1.59.0
requests>=2.28.0
PyYAML>=6.0
pandas>=1.4.0
tqdm>=4.62.0
Pillow>=9.0.0
